Fix modal not getting closed when page goes inactive
