+++ src/components/OpenAppFailureModal/index.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import BaseOpenAppFailureModal from './BaseOpenAppFailureModal';
+
+/** Reloads the app to trigger OpenApp reconnection */
+const reloadApp = () => {
+    window.location.reload();
+};
+
+function OpenAppFailureModal() {
+    return <BaseOpenAppFailureModal onRefreshAndTryAgainButtonPress={reloadApp} />;
+}
+
+OpenAppFailureModal.displayName = 'OpenAppFailureModal';
+
+export default OpenAppFailureModal;
