--- src/hooks/useOpenAppReset/index.native.ts
@@ -1,9 +1,0 @@
-import {openApp} from '@libs/actions/App';
-
-const useOpenAppReset = () => {
-    return () => {
-        openApp();
-    };
-};
-
-export default useOpenAppReset;
