--- src/hooks/useOpenAppReset/index.ts
@@ -1,7 +1,0 @@
-const useOpenAppReset = () => {
-    return () => {
-        window.location.reload();
-    };
-};
-
-export default useOpenAppReset;
